// src/store/mian/system/type.ts
export interface systemState {
  userList: [],
  userListTotalCount: number
  department: [],
  departmentTotalCount: number
  roleList: [],
  roleListTotalCount: number
  menuList: [],
  menuListTotalCount: number
  categoryList: [],
  categoryListTotalCount: number
  goodsList: [],
  goodsListTotalCount: number
}